Adjusted the routes direction and seperated the login and signup page into diferent files
