+++ src/components/modules/TogglePassword.js
@@ -1,0 +1,9 @@
+function TogglePassword(password, toggle){
+    toggle.addEventListener('click', () => {
+        if(password.type === 'password'){
+            password.type = 'text'
+        }else{
+            password.type = 'password';
+        }
+    })
+}
