+++ src/styles/login.scss
@@ -2,16 +2,46 @@ $poppins: "Poppins", sans-serif;
 $pink: #FACBCD;
 $lato: "Lato", sans-serif;
 $red: #D22D33;
-body{
+.main-signup{
     background-color: $pink;
     overflow-x: hidden;
+    .floating-img-1{
+        position: absolute;
+        left: -301px;
+        top: 401px;
+        z-index: 1;
+        img{
+            width: 503px;
+        }
+    }
+    .floating-img-2{
+        position: absolute;
+        right: -50px;
+        top: 1040px;
+        z-index: 1;
+        img{
+            width: 204px;
+        }
+    }
+    .floating-img-3{
+        position: absolute;
+        top: -160px;
+        right: 50px;
+        z-index: 1;
+        img{
+            width: 435px;
+        }
+    }
 }
-main{
+.main-body{
     margin: 0 auto;
     margin-top: 50px;
     display: flex;
     flex-direction: column;
     position: relative;
+    *{
+        z-index: 10;
+    }
     .quickchow-logo{
         margin-bottom: 30px;
         margin-left: 150px;
@@ -34,70 +64,56 @@ main{
             line-height: 30px;
             font-family: $poppins;
         }
-        label{
-            font-family: $poppins;
-            display: block;
-        }
-        .input-elem{
-            width: 100%;
-            height: 70px;
-            border-radius: 10px;
-            background-color: #f4f4f4;
-            border: none;
-            margin-top: 5px;
-            padding-left: 20px;
-            outline: none;
-            .selector-span{
-                color: red !important;
-                font-size: 2rem;
-            }
-        }
-        .input-elem:focus{
-            border: 1px solid $red;
-        }
         .user-input{
             display: flex;
             flex-direction: column;
             gap: 40px;
             margin-top: 50px;
-        }
-        .isStudent{
-            h2{
-                font-size: 1.2rem;
-                margin-bottom: 20px;
+            .input-elem{
+                width: 100%;
+                height: 70px;
+                border-radius: 10px;
+                background-color: #f4f4f4;
+                border: none;
+                margin-top: 5px;
+                padding-left: 20px;
+                outline: none;
+                .selector-span{
+                    color: red !important;
+                    font-size: 2rem;
+                }
             }
-            div{
-                display: flex;
-                gap: 20px;
+            .input-elem:focus{
+                border: 1px solid $red;
             }
-            input{
-                margin-right: 5px;
+            label{
+                font-family: $poppins;
+                display: block;
             }
-        }
-    }
-    .floating-img-1{
-        position: absolute;
-        left: -301px;
-        top: 401px;
-        img{
-            width: 503px;
-        }
-    }
-    .floating-img-2{
-        position: absolute;
-        right: -50px;
-        top: 1040px;
-        img{
-            width: 204px;
-        }
-    }
-    .floating-img-3{
-        position: absolute;
-        top: -249px;
-        right: 50px;
-        z-index: -1;
-        img{
-            width: 435px;
+            .isStudent{
+                h2{
+                    font-size: 1.2rem;
+                    margin-bottom: 20px;
+                }
+                div{
+                    display: flex;
+                    gap: 20px;
+                }
+                input{
+                    margin-right: 5px;
+                }
+            }
+            .password-label{
+                position: relative;
+                img{
+                    position: absolute;
+                    bottom: 20px;
+                    right: 20px;
+                    width: 24px;
+                    cursor: pointer;
+                    // filter: invert(28%) sepia(63%) saturate(2564%) hue-rotate(337deg) brightness(83%) contrast(99%);
+                }
+            }
         }
     }
     .sign-up{
